py/adk-web/requirements.txt
# ADK and Braintrust dependencies
braintrust-adk==0.3.1
braintrust==0.5.0
google-adk==1.22.1

# Google GenAI dependencies
google-genai==1.60.0
google-api-core==2.29.0

# Web server
fastapi==0.123.10
uvicorn==0.40.0
starlette==0.50.0
httpx==0.28.1
httpx-sse==0.4.3
websockets==15.0.1
sse-starlette==3.2.0

# Environment management
python-dotenv==1.2.1

# Core dependencies
pydantic==2.12.5
typing_extensions==4.15.0
python-multipart==0.0.21
